Moving indentical functions to a single module.
