+++ vsg/rules/multiline_procedure_call_structure.py
@@ -4,6 +4,7 @@ from vsg import token
 from vsg import violation
 
 from vsg.rules import utils as rules_utils
+from vsg.rules import check
 from vsg.rules import create_violation
 from vsg.rules import fix
 from vsg.rules import tokens_of_interest as toi
@@ -51,47 +52,22 @@ class multiline_procedure_call_structure(structure.Rule):
 
 def _check_first_open_paren(self, oToi):
 
-    _check_add_new_line_and_remove_new_line(self, oToi, self.first_open_paren, token.procedure_call.open_parenthesis)
+    check.add_new_line_and_remove_new_line(self, oToi, self.first_open_paren, token.procedure_call.open_parenthesis)
 
 
 def _check_last_close_paren(self, oToi):
 
-    _check_add_new_line_and_remove_new_line(self, oToi, self.last_close_paren, token.procedure_call.close_parenthesis)
+    check.add_new_line_and_remove_new_line(self, oToi, self.last_close_paren, token.procedure_call.close_parenthesis)
 
 
 def _check_association_list_comma(self, oToi):
 
-    _check_add_new_line_and_remove_new_line(self, oToi, self.association_list_comma, token.association_list.comma)
+    check.add_new_line_and_remove_new_line(self, oToi, self.association_list_comma, token.association_list.comma)
 
 
 def _check_association_element(self, oToi):
 
-    _check_add_new_line_and_remove_new_line(self, oToi, self.association_element, token.association_element.formal_part)
-
-
-def _check_add_new_line_and_remove_new_line(self, oToi, sOption, oTokenType):
-    if sOption == 'ignore':
-        return
-    elif sOption == 'add_new_line':
-        rules_utils.analyze_with_function(self, oToi, oTokenType, analyze_add_new_line)
-    elif sOption == 'remove_new_line':
-        rules_utils.analyze_with_function(self, oToi, oTokenType, analyze_remove_new_line)
-
-
-def analyze_add_new_line(self, oToi):
-    iToken = oToi.get_meta_data('iToken')
-    lTokens = oToi.get_tokens()
-    if not rules_utils.token_at_beginning_of_line_in_token_list(iToken, lTokens):
-        oViolation = create_violation.add_new_line(oToi)
-        self.add_violation(oViolation)
-
-
-def analyze_remove_new_line(self, oToi):
-    iToken = oToi.get_meta_data('iToken')
-    lTokens = oToi.get_tokens()
-    if rules_utils.token_at_beginning_of_line_in_token_list(iToken, lTokens):
-        oViolation = create_violation.remove_new_line(self, oToi)
-        self.add_violation(oViolation)
+    check.add_new_line_and_remove_new_line(self, oToi, self.association_element, token.association_element.formal_part)
 
 
 def first_open_paren_detected(oToi, iLine, oToken):
